Update post.go and update.go methods, bump version
The UpdateCall method in update.go has been changed to take a string instead of a byte slice and the UpdateCall call in post.go has been adjusted to match. A new hashing function has been added to generate a unique identifier for each post, which is then included in the Location header of the server's response. Furthermore, the version has been bumped from 0.0.5 to 0.0.6 and the compose.yaml file has been edited to better suit the new version. Documentation has also been extended.